cricket_analytics/requirement_file.txt
appnope==0.1.0
astroid==1.6.6
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
cffi==1.14.3
configparser==4.0.2
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.2
Django==1.8
ecdsa==0.14.1
enum34==1.1.10
futures==3.3.0
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.10.0
ipython-genutils==0.2.0
isort==4.3.21
jupyter-client==5.3.5
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.5.1
matplotlib==2.2.5
mccabe==0.6.1
numpy==1.16.6
pandas==0.24.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
Pygments==2.5.2
pylint==1.9.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.1
pyzmq==19.0.2
rsa==4.5
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.3
seaborn==0.9.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
subprocess32==3.5.4
tornado==5.1.1
traitlets==4.3.3
wcwidth==0.2.5
wrapt==1.12.1
